testing/web-platform/tests/service-workers/service-worker/local-url-inherit-controller.https.html

<!DOCTYPE html>
<title>Service Worker: local URL windows and workers inherit controller</title>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

const SCRIPT = 'resources/empty.js';
const SCOPE = 'resources/local-url-inherit-controller-frame.html';

async function doAsyncTest(t, opts) {
  let name = opts.scheme + '-' + opts.child;
  let scope = SCOPE + '?name=' + name;
  let reg = await service_worker_unregister_and_register(t, SCRIPT, scope);
  add_completion_callback(_ => reg.unregister());
  await wait_for_state(t, reg.installing, 'activated');

  let frame = await with_iframe(scope);
  add_completion_callback(_ => frame.remove());
  assert_not_equals(frame.contentWindow.navigator.serviceWorker.controller, null,
                    'frame should be controlled');

  let blobController = await frame.contentWindow.checkChildController(opts);

  let expect = opts.expect === 'inherit'
             ?  frame.contentWindow.navigator.serviceWorker.controller.scriptURL
             : null;

  assert_equals(blobController, expect,
                `${opts.scheme} URL ${opts.child} should ${opts.expect} controller`);
}

promise_test(function(t) {
  return doAsyncTest(t, {
    scheme: 'blob',
    child: 'iframe',
    expect: 'inherit',
  });
}, 'Same-origin blob URL iframe should inherit service worker controller.');

promise_test(function(t) {
  return doAsyncTest(t, {
    scheme: 'blob',
    child: 'worker',
    expect: 'inherit',
  });
}, 'Same-origin blob URL worker should inherit service worker controller.');

promise_test(function(t) {
  // Data URLs should result in an opaque origin and should probably not
  // have access to a cross-origin service worker.  See:
  //
  // https://github.com/w3c/ServiceWorker/issues/1262
  //
  return doAsyncTest(t, {
    scheme: 'data',
    child: 'iframe',
    expect: 'not inherit',
  });
}, 'Data URL iframe should not inherit service worker controller.');

promise_test(function(t) {
  // Data URLs should result in an opaque origin and should probably not
  // have access to a cross-origin service worker.  See:
  //
  // https://github.com/w3c/ServiceWorker/issues/1262
  //
  return doAsyncTest(t, {
    scheme: 'data',
    child: 'worker',
    expect: 'not inherit',
  });
}, 'Data URL worker should not inherit service worker controller.');

</script>
</body>
